+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.30](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.29...@stream-io/video-react-bindings-0.2.30) (2023-10-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.3.29`
 ### [0.2.29](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.28...@stream-io/video-react-bindings-0.2.29) (2023-09-28)
 
 ### Dependency Updates